feat: change `workout/[dateId]` and `workout/[dateId]/sets`
